+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add 'register' command. Allows users to register their Destiny 2 account with the bot
 - The roster will now display a member's nickname instead of their username if a nickname is set
 - Add error message when invoking a command that a user doesn't have permissions to invoke
+- Add owner only command to view current bot stats
 ### Changed
 - Reactions are no longer auto removed from events. This is to cut down on the number of API calls
 the bot has to make